fix: knit the rebase, and act on the landing review
The rebase integration: master's rename cascade learns the referrers
the feature added — a mesh's skeleton path, a clip set's, a .bvat's
three inputs (rewritten with stamps intact, pinned by a rig-rename
test); a skeleton stores no references and says so; master's new tests
come along to the feature's APIs (loadMeshRefs, core::string_pool);
vatData/time/prevTime join ForwardPass's binder-name validation.

The landing review: scoped ChunkId enums across all four containers
(chunk::Writer/Reader take any uint32-backed enum by concept); .bmesh
drops its cosmetic minor bump; magics spell their tags via constexpr
fourCC; findBone returns optional; the glTF skin names its sampler
indices; the clock pair is DrawData::TimeData (draw.clock); the VAT
module is Forward_VatMesh, completing the family naming; Get-prefixed
editor getters; the bone-over-joint rule and the tangent story written
into the docs.

+++ libs/assetlib_structs/include/assetlib_structs/VertexLayout.h
@@ -64,7 +64,7 @@ namespace assetlib
 	 * Influences a skinned vertex carries: `kJoints0` and `kWeights0` are a vec4 pair, so both the
 	 * importer that writes them and anything that decodes them work in fours.
 	 */
-	inline constexpr size_t c_InfluencesPerVertex = 4;
+	inline constexpr uint32_t c_InfluencesPerVertex = 4;
 
 	/**
 	 * Byte offset of `semantic` within one interleaved vertex, empty when the layout does not carry
